requirements/base.txt
algoliasearch-django==1.5.1
argon2-cffi==16.3.0
awesome-slugify==1.6.5
boto3==1.4.7
celery==3.1.25
django==1.10.8 # pyup: >=1.10,<1.11
django-allauth==0.33.0
django-anymail==1.0
django-crispy-forms==1.6.1
django-environ==0.4.4
django-meta==1.4
django-model-utils==3.0.0
django-redis==4.8.0
django-storages==1.6.5
django-taggit==0.22.2
Fabric3==1.13.1.post1
gevent==1.2.2
gunicorn==19.7.1
mailchimp3==3.0.1
ipdb==0.10.3
ipython==6.1.0
Pillow==4.2.1
pip-tools==1.9.0
psycopg2==2.7.3.1
pytz==2017.2
redis>=2.10.5
sorl-thumbnail==12.4.1
wheel==0.30.0
